refactor(pms): 调整权限与角色相关请求参数类型及配置
- 将 AddRolePermissionsRequest 中的 id 和 permissionIds 类型从 Long 改为 Integer- 将 AddRoleUsersRequest 中的 userIds 类型从 Long 改为 Integer
- 将 AddUserRolesRequest 中的 roleIds 类型从 Long 改为 Integer
- 修改 application.yml 默认激活配置为 mysql
- 更新 application-mysql.yml 数据源配置结构和连接池参数
- 移除 AuthController 中的 Swagger 注解及相关验证码逻辑
- 删除 BCryptTest 测试类
- 简化 CaptchaServiceImpl 中的方法注解
- 移除 ChangePasswordRequest 中的非空校验注解
- 删除 checkstyle 相关配置文件
- 新增全局异常处理类 ControllerExceptionAdvice- 添加类型转换工具类 ConvertUtil
- 修改 CreatePermissionRequest 和 CreateRoleRequest 请求类实现自定义 convert 方法并移除 mapstruct 依赖- 新增 mybatis generator 配置文件 generatorConfig.xml
- 添加数据库初始化脚本 docs/init.sql

+++ src/main/java/cn/dhbin/isme/pms/domain/request/AddRolePermissionsRequest.java
@@ -11,7 +11,7 @@ import lombok.Data;
 @Data
 public class AddRolePermissionsRequest {
 
-    private Long id;
+    private Integer id;
 
-    private List<Long> permissionIds;
+    private List<Integer> permissionIds;
 }
