+++ src/main/java/cn/dhbin/isme/pms/domain/request/RemoveRoleUsersRequest.java
@@ -11,6 +11,6 @@ import lombok.Data;
 @Data
 public class RemoveRoleUsersRequest {
 
-    private List<Long> userIds;
+    private List<Integer> userIds;
 
 }
